Refactor new-tab WebUI
BasicUI had some things only specific to the new-tab in it.
This change makes it so the custom things for NewTabUI are in the NewTab
page and makes it easier for future things to subclass BasicUI
