+++ browser/ui/BUILD.gn
@@ -23,12 +23,14 @@ source_set("ui") {
     "views/importer/brave_import_lock_dialog_view.h",
     "webui/basic_ui.cc",
     "webui/basic_ui.h",
+    "webui/brave_new_tab_ui.cc",
+    "webui/brave_new_tab_ui.h",
     "webui/brave_web_ui_controller_factory.cc",
     "webui/brave_web_ui_controller_factory.h",
     "webui/brave_welcome_ui.cc",
     "webui/brave_welcome_ui.h",
-    "webui/new_tab_html_source.cc",
-    "webui/new_tab_html_source.h",
+    "webui/brave_webui_source.cc",
+    "webui/brave_webui_source.h",
   ]
 
   if (is_mac && !mac_views_browser) {
