+++ browser/ui/webui/basic_ui.cc
@@ -4,21 +4,11 @@
 
 #include "brave/browser/ui/webui/basic_ui.h"
 
-#include "brave/browser/ui/webui/new_tab_html_source.h"
-#include "brave/common/pref_names.h"
-#include "brave/common/webui_url_constants.h"
+#include "brave/browser/ui/webui/brave_webui_source.h"
 #include "chrome/browser/profiles/profile.h"
-#include "chrome/common/url_constants.h"
-#include "components/prefs/pref_change_registrar.h"
-#include "components/prefs/pref_service.h"
-#include "content/public/browser/render_frame_host.h"
 #include "content/public/browser/web_ui_data_source.h"
 #include "content/public/browser/web_ui_message_handler.h"
-#include "content/public/common/bindings_policy.h"
-#include "ui/base/resource/resource_bundle.h"
 
-
-using content::WebContents;
 using content::WebUIMessageHandler;
 
 namespace {
@@ -30,12 +20,10 @@ content::WebUIDataSource* CreateBasicUIHTMLSource(Profile* profile,
                                                   int html_resource_id) {
   content::WebUIDataSource* source =
       content::WebUIDataSource::Create(name);
-
-  CustomizeNewTabHTMLSource(source);
-
   source->SetJsonPath("strings.js");
   source->SetDefaultResource(html_resource_id);
   source->AddResourcePath(js_file, js_resource_id);
+  CustomizeWebUIHTMLSource(name, source);
   return source;
 }
 
@@ -74,34 +62,10 @@ BasicUI::BasicUI(content::WebUI* web_ui,
   BasicDOMHandler* handler = handler_owner.get();
   web_ui->AddMessageHandler(std::move(handler_owner));
   handler->Init();
-  content::WebUIDataSource::Add(profile,
-    CreateBasicUIHTMLSource(profile, name,
-      js_file, js_resource_id, html_resource_id));
-  PrefService* prefs = profile->GetPrefs();
-  pref_change_registrar_ = std::make_unique<PrefChangeRegistrar>();
-  pref_change_registrar_->Init(prefs);
-  pref_change_registrar_->Add(kAdsBlocked,
-    base::Bind(&BasicUI::OnPreferenceChanged, base::Unretained(this)));
-  pref_change_registrar_->Add(kTrackersBlocked,
-    base::Bind(&BasicUI::OnPreferenceChanged, base::Unretained(this)));
-  pref_change_registrar_->Add(kHttpsUpgrades,
-    base::Bind(&BasicUI::OnPreferenceChanged, base::Unretained(this)));
+  content::WebUIDataSource* source = CreateBasicUIHTMLSource(profile, name,
+      js_file, js_resource_id, html_resource_id);
+  content::WebUIDataSource::Add(profile, source);
 }
 
 BasicUI::~BasicUI() {
-  pref_change_registrar_.reset();
 }
-
-void BasicUI::RenderFrameCreated(content::RenderFrameHost* render_frame_host) {
-  if (0 != (web_ui()->GetBindings() & content::BINDINGS_POLICY_WEB_UI)) {
-    CustomizeNewTabWebUIProperties(web_ui());
-  }
-}
-
-void BasicUI::OnPreferenceChanged() {
-  if (0 != (web_ui()->GetBindings() & content::BINDINGS_POLICY_WEB_UI)) {
-    CustomizeNewTabWebUIProperties(web_ui());
-    web_ui()->CallJavascriptFunctionUnsafe("brave_new_tab.statsUpdated");
-  }
-}
-
