+++ browser/ui/webui/brave_webui_source.h
@@ -2,11 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+#include <string>
+
 namespace content {
 class WebUIDataSource;
 class WebUI;
 }
 
-
-void CustomizeNewTabHTMLSource(content::WebUIDataSource* source);
-void CustomizeNewTabWebUIProperties(content::WebUI* web_ui);
+void CustomizeWebUIHTMLSource(const std::string &name, content::WebUIDataSource* source);
